project crud

+++ src/main/java/uz/elmurodov/spring_boot/services/project/ProjectServiceImpl.java
@@ -3,11 +3,11 @@ package uz.elmurodov.spring_boot.services.project;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uz.elmurodov.spring_boot.criteria.GenericCriteria;
+import uz.elmurodov.spring_boot.dto.project.ProjectColumnCreateDto;
 import uz.elmurodov.spring_boot.dto.project.ProjectCreateDto;
 import uz.elmurodov.spring_boot.dto.project.ProjectDto;
 import uz.elmurodov.spring_boot.dto.project.ProjectUpdateDto;
 import uz.elmurodov.spring_boot.entity.base.AuditAwareImpl;
-import uz.elmurodov.spring_boot.entity.organization.Organization;
 import uz.elmurodov.spring_boot.entity.project.Project;
 import uz.elmurodov.spring_boot.mapper.project.ProjectMapper;
 import uz.elmurodov.spring_boot.reposiroty.project.ProjectRepository;
@@ -27,43 +27,49 @@ public class ProjectServiceImpl extends AbstractService<
         ProjectValidator> implements ProjectService {
 
     private final AuditAwareImpl auditAware;
+
     @Autowired
     protected ProjectServiceImpl(
             ProjectRepository repository,
             ProjectMapper mapper,
             ProjectValidator validator,
-            BaseUtils baseUtils, AuditAwareImpl auditAware) {
+            BaseUtils baseUtils, AuditAwareImpl auditAware, ProjectColumnService projectColumnService) {
         super(repository, mapper, validator, baseUtils);
         this.auditAware = auditAware;
     }
 
     @Override
     public Long create(ProjectCreateDto createDto) {
+        if (!auditAware.getCredentials().getRole().getCode().equals("ADMIN")) throw new RuntimeException("403");
         Project project = mapper.fromCreateDto(createDto);
         project.setOrganization(auditAware.getCredentials().getOrganizationId());
         repository.save(project);
         return project.getId();
     }
 
+
     @Override
     public Void delete(Long id) {
+        if (!auditAware.getCredentials().getRole().getCode().equals("ADMIN")) throw new RuntimeException("403");
         repository.deleteSoft(id);
         return null;
     }
 
     @Override
     public Void update(ProjectUpdateDto updateDto) {
+        if (!auditAware.getCredentials().getRole().getCode().equals("ADMIN")) throw new RuntimeException("403");
+        repository.update(updateDto);
         return null;
     }
 
     @Override
     public List<ProjectDto> getAll(GenericCriteria criteria) {
-        return null;
+        return mapper.toDto(repository.getAll());
     }
 
     @Override
     public ProjectDto get(Long id) {
-        return null;
+        return mapper.toDto(repository.getProject(id));
     }
 
     @Override
